+++ saas/.importlinter
@@ -1,0 +1,35 @@
+[importlinter]
+root_package = backend
+include_external_packages = True
+
+[importlinter:contract:layers-apps-service-biz]
+name=apps service biz component
+type=layers
+layers =
+    backend.apps
+    backend.trans
+    backend.biz
+    backend.service
+    backend.component
+ignore_imports =
+    backend.trans.* -> backend.apps.*.models
+    backend.biz.* -> backend.apps.*.models
+    backend.biz.*.* -> backend.apps.*.models
+    backend.service.* -> backend.apps.*.models
+    backend.service.*.* -> backend.apps.*.models
+    backend.biz.org_sync.syncer -> backend.apps.organization.tasks
+    backend.biz.application -> backend.apps.role.tasks
+    backend.biz.application -> backend.audit.audit
+    backend.biz.handover -> backend.audit.audit
+
+[importlinter:contract:layers-api-service-biz]
+name=api service biz component
+type=layers
+layers =
+    backend.api
+    backend.trans
+    backend.biz
+    backend.service
+    backend.component
+ignore_imports =
+    backend.biz.model_event -> backend.api.authorization.models
