+++ saas/.pre-commit-config.yaml
@@ -36,6 +36,11 @@ repos:
         language: python
         pass_filenames: false
         entry: mypy --config-file=saas/pyproject.toml saas
+      - id: import-linter
+        name: import-linter
+        language: python
+        pass_filenames: false
+        entry: cd saas && lint-imports --config=./.importlinter && cd ..
       - id: pytest
         name: pytest
         language: python
