+++ saas/pyproject.toml
@@ -153,6 +153,7 @@ isort = "^5.9.2"
 pytest-cov = "^3.0.0"
 flake8-bugbear = "22.9.23"
 bandit = "1.7.1"
+import-linter = "1.2.6"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
