+++ saas/requirements_dev.txt
@@ -26,7 +26,7 @@ charset-normalizer==2.0.4; python_full_version >= "3.6.2" and python_version >= 
 click-didyoumean==0.3.0; python_full_version >= "3.6.2" and python_full_version < "4.0.0" and python_version >= "3.6"
 click-plugins==1.1.1; python_version >= "3.6"
 click-repl==0.2.0; python_version >= "3.6"
-click==7.1.2; python_full_version >= "3.6.2" and python_version >= "3.6" and python_version < "3.11" and python_full_version < "4.0.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6")
+click==7.1.2; python_full_version >= "3.6.2" and python_version >= "3.6" and python_version < "3.11" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6") and python_full_version < "4.0.0"
 colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0" and platform_system == "Windows"
 contextvars==2.4; python_version < "3.7" and python_version >= "3.6"
 converge==0.9.8
@@ -36,7 +36,7 @@ coverage==6.2; python_version >= "3.6"
 cryptography==3.3.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
 curlify==2.2.1; python_full_version >= "3.6.2" and python_version < "3.11" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4") and python_full_version < "4.0.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0")
 dataclasses==0.7; python_version >= "3.6" and python_version < "3.7"
-decorator==5.0.9; python_version >= "3.5"
+decorator==5.0.9; python_version >= "3.6"
 deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
 django-celery-beat==2.2.1
 django-cors-headers==3.4.0; python_version >= "3.5"
@@ -62,11 +62,13 @@ gitdb==4.0.9; python_version >= "3.6"
 gitpython==3.1.20; python_version >= "3.6"
 googleapis-common-protos==1.54.0; python_version >= "3.6"
 greenlet==1.1.2; python_version >= "2.7" and python_full_version < "3.0.0" and platform_python_implementation == "CPython" or python_version > "3.5" and python_full_version < "3.0.0" and platform_python_implementation == "CPython" or python_version > "3.5" and platform_python_implementation == "CPython" and python_full_version >= "3.5.0"
+grimp==1.3; python_version >= "3.6"
 grpcio==1.42.0; python_version >= "3.6"
 gunicorn==20.1.0; python_version >= "3.5"
 httplib2==0.19.0
 idna==2.10; python_full_version >= "3.6.2" and python_version >= "3" and python_version < "3.11" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4") and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") and python_full_version < "4.0.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0")
 immutables==0.16; python_version < "3.7" and python_version >= "3.6"
+import-linter==1.2.6; python_version >= "3.6"
 importlib-metadata==4.8.3; python_version >= "3.6"
 inflection==0.5.1; python_version >= "3.6"
 iniconfig==1.1.1; python_version >= "3.6"
@@ -80,6 +82,7 @@ mock==1.0.1
 mypy-extensions==0.4.3; python_full_version >= "3.6.2" and python_version >= "3.5"
 mypy==0.910; python_version >= "3.5"
 mysqlclient==2.0.1; python_version >= "3.5"
+networkx==2.5; python_version >= "3.6"
 openpyxl==3.0.9; python_version >= "3.6"
 opentelemetry-api==1.7.1; python_version >= "3.6"
 opentelemetry-exporter-jaeger-proto-grpc==1.7.1; python_version >= "3.6"
